requirements.txt
Flask
flask-cors
Flask-Migrate
Flask-SQLAlchemy
flask-restful
psycopg2-binary
python-dotenv

